% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCytoInfo.R
\name{addCytoInfo}
\alias{addCytoInfo}
\title{omnibus sub-chromosome splitting information (chr, arm, and band)}
\usage{
addCytoInfo(x, y = c("chr", "arm", "band"))
}
\arguments{
\item{x}{a SummarizedExperiment-derived object or something like it}

\item{y}{what kind of information to add (chr, arm, band)}
}
\value{
x, but with y added to mcols(x)
}
\description{
omnibus sub-chromosome splitting information (chr, arm, and band)
}

