% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asTDF.R
\name{asTDF}
\alias{asTDF}
\title{Dump DNA methylation values, gene-level counts, or whatever for IGV viewing}
\usage{
asTDF(x, assay = 1, group = NULL, stub = NULL, genome = "hg19")
}
\arguments{
\item{x}{a SummarizedExperiment-like object of some sort}

\item{assay}{the name of the assay to dump (default: assays(x)[[1]])}

\item{group}{a grouping factor to dump by (default: dump everyone)}

\item{stub}{the filename stub to dump (default: name of x, or x_group)}

\item{genome}{the genome to use (default is hg19, due to annotations)}
}
\value{
list     a list of filenames dumped
}
\description{
Dump DNA methylation values, gene-level counts, or whatever for IGV viewing
}

