% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToMultiAssay.R
\name{convertToMultiAssay}
\alias{convertToMultiAssay}
\title{Take a RangedSummarizedExperiment x & turn it into a MultiAssayExperiment y
where the former metadata(x) becomes the integrity-checked linkedAssays(y).}
\usage{
convertToMultiAssay(x)
}
\arguments{
\item{x}{a RangedSummarizedExperiment whose metadata() contains assays}
}
\value{
a MultiAssayExperiment y with integrity-checked linkedAssays(y)
}
\description{
Take a RangedSummarizedExperiment x & turn it into a MultiAssayExperiment y
where the former metadata(x) becomes the integrity-checked linkedAssays(y).
}

