% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPredictors.R
\name{createPredictors}
\alias{createPredictors}
\title{as the title suggests.  Original author Zhang (cf. Genetic Epi, 2016)}
\usage{
createPredictors(covariates = NULL, funcs, kz = 30, kb = 30,
  smooth.cov = FALSE)
}
\description{
as the title suggests.  Original author Zhang (cf. Genetic Epi, 2016)
}

