% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch450k.R
\name{fetch450k}
\alias{fetch450k}
\title{Pull a 450k dataset from GEO.}
\usage{
fetch450k(GSE)
}
\arguments{
\item{GSE}{The dataset.}
}
\value{
a GenomicRatioSet.
}
\description{
Doesn't matter what the annotations are or any other bullshit. Just get it.
}

