% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchEPIC.R
\name{fetchEPIC}
\alias{fetchEPIC}
\title{Pull an EPIC dataset from GEO.}
\usage{
fetchEPIC(GSE, ...)
}
\arguments{
\item{GSE}{The dataset.}

\item{...}{Additional arguments to be passed to prepForReanalysis().}
}
\value{
a GenomicRatioSet.
}
\description{
Doesn't matter what the annotations are or any other bullshit. Just get it.
}

