% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitMoM.R
\name{fitMoM}
\alias{fitMoM}
\title{function to fit Mas-o-Menos (MoM) predictor scores (usually) against DMRs}
\usage{
fitMoM(atDMRs, tumorName = NULL)
}
\arguments{
\item{atDMRs}{SummarizedExperiment-derived object collapsed across regions}

\item{tumorName}{Optional tumor name for the plots (default: guess from args)}
}
\value{
a list with K-M and Cox fits, effect sizes & mm coefficients
}
\description{
function to fit Mas-o-Menos (MoM) predictor scores (usually) against DMRs
}

