% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIDATsToRgSet.R
\name{geoIDATsToRgSet}
\alias{geoIDATsToRgSet}
\title{short-circuit minfi's irritating insistence on un-gzipped IDAT files}
\usage{
geoIDATsToRgSet(res)
}
\arguments{
\item{res}{something with $supplementary_file and $supplementary_file.1}
}
\value{
an rgSet
}
\description{
short-circuit minfi's irritating insistence on un-gzipped IDAT files
}

