% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDATsToRgSet.R
\name{IDATsToRgSet}
\alias{IDATsToRgSet}
\title{Function to sidestep minfi's irritating insistence on un-gzipped IDAT files}
\usage{
IDATsToRgSet(res, idatGrn = NULL, idatRed = NULL, parallel = FALSE)
}
\arguments{
\item{res}{a data.frame with c(idatGrn, idatRed) OR Basename column(s)}

\item{idatGrn}{the column with the green IDAT filenames, if any}

\item{idatRed}{the column with the red IDAT filenames, if any}

\item{parallel}{try to run in parallel? (default: FALSE)}
}
\value{
an rgSet
}
\description{
Function to sidestep minfi's irritating insistence on un-gzipped IDAT files
}

