% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedExperiment-class.R, R/LinkedExperiment-methods.R
\docType{class}
\name{LinkedExperiment-class}
\alias{LinkedExperiment-class}
\alias{[,LinkedExperiment,ANY,ANY,ANY-method}
\title{A master container for linked experiments.  The idea is that the parent has 
no assays, and the children have no covariates.  So initializing one of these
would usually occur by "lifting" the colData out of a SummarizedExperiment
and into the LinkedExperiments, then "demoting" everything else about it 
into the child containers.  I think the validity method will end up checking
that assays() is empty (heh) and that all child assays have colnames like 
the parent's.  This could be relaxed somehow, but I'm not sure how just yet.}
\usage{
\S4method{[}{LinkedExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a LinkedExperiment}

\item{i}{optional row index or indices (will be ignored)}

\item{j}{optional column index or indices (will be used)}
}
\value{
A valid LinkedExperiment w/column-matching linked assays
}
\description{
A master container for linked experiments.  The idea is that the parent has 
no assays, and the children have no covariates.  So initializing one of these
would usually occur by "lifting" the colData out of a SummarizedExperiment
and into the LinkedExperiments, then "demoting" everything else about it 
into the child containers.  I think the validity method will end up checking
that assays() is empty (heh) and that all child assays have colnames like 
the parent's.  This could be relaxed somehow, but I'm not sure how just yet.
}
\details{
If and only if 'j' is set, the LinkedExperiment 'x' 
               briefly becomes a 'RangedSummarizedExperiment' while its 
               linkedAssays are recursively subsetted. The two pieces are 
               then put back together as a valid LinkedExperiment.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: A referential-integrity-preserving subset operator
}}
\section{Slots}{

\describe{
\item{\code{linkedAssays}}{multiple assays, presumably for each sample (or add NAs)}
}}

