% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map27k.R
\name{map27k}
\alias{map27k}
\title{wrapper for preprocessNoob27k}
\usage{
map27k(mset, genome = "hg19", ...)
}
\arguments{
\item{genome}{a genome assembly (currently only hg19 works properly)}

\item{...}{additional arguments to pass to}

\item{rgSet}{a humanMethylation27 rgSet}
}
\value{
a GenomicRatioSet
}
\description{
wrapper for preprocessNoob27k
}

