% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{MultiAssayExperiment-class}
\title{Something like a SummarizedExperiment, but with "holes" and metadata checks}
\description{
Something like a SummarizedExperiment, but with "holes" and metadata checks
}
\section{Slots}{

\describe{
\item{\code{perSampleMetadata}}{associated per-sample data, checked against colnames}
}}

