% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old2new.R
\name{old2new}
\alias{old2new}
\title{to update old grSets so I don't have to reload, e.g. cnAML
note: this can be a pig if the dataset is huge; in that case, just reload it}
\usage{
old2new(grSet)
}
\arguments{
\item{grSet}{the old grSet}
}
\value{
a new, valid grSet
}
\description{
to update old grSets so I don't have to reload, e.g. cnAML
note: this can be a pig if the dataset is huge; in that case, just reload it
}

