% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{stripMcols}
\alias{EFSurv}
\alias{OSurv}
\alias{getSurv}
\alias{hasSurv}
\alias{maskBetas}
\alias{stripMcols}
\title{for tidying up rowRanges prior to merging}
\usage{
stripMcols(x)

maskBetas(grSet, pcutoff = 0.01, ...)

hasSurv(x, time = "time", event = "event")

getSurv(x, time = "time", event = "event")

OSurv(x, time = "OS", event = "OSevent")

EFSurv(x, time = "EFS", event = "EFSevent")
}
\arguments{
\item{x}{a GRanges or GRangesList that has metadata columns it doesn't need}

\item{grSet}{a GenomicRatioSet with matrices "Beta" and "pval"}

\item{pcutoff}{optional p-value cutoff for masking, default is 0.01}

\item{time}{column name for time to event covariate, default "time"}

\item{event}{column name for censoring/event covariate, default "event"}

\item{x}{a SummarizedExperiment-like dataset}

\item{x}{a SummarizedExperiment-like dataset}

\item{time}{column name for time to event, default "OS"}

\item{event}{column name for censoring/event (0/1), default "OSevent"}

\item{x}{a SummarizedExperiment-like dataset}

\item{time}{column name for time to event, default "OS"}

\item{event}{column name for censoring/event (0/1), default "OSevent"}

\item{x}{a SummarizedExperiment-like dataset}

\item{time}{column name for time to event, default "EFS"}

\item{event}{column name for censoring/event (0/1), default "EFSevent"}
}
\value{
a boolean vector: whether each column has both covariates

a \code{Surv} object, if one can be constructed from x

a \code{Surv} object, if one can be constructed from x

a \code{Surv} object, if one can be constructed from x
}
\description{
for tidying up rowRanges prior to merging

mask (set to NA) beta values by pval for a GenomicRatioSet

convenience function to feed getSurv, OSurv, EFSurv

convenience function to feed OSurv, EFSurv

Get overall survival for a SummarizedExperiment-like dataset

Get event-free survival from SummarizedExperiment-like data
}
\section{Functions}{
\itemize{
\item \code{stripMcols}: 

\item \code{maskBetas}: 

\item \code{hasSurv}: 

\item \code{getSurv}: 

\item \code{OSurv}: 

\item \code{EFSurv}: 
}}

