% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSNPs.R
\name{plotSNPs}
\alias{plotSNPs}
\title{plot SNP probes from DNA methylation arrays to avoid label swaps}
\usage{
plotSNPs(x, individuals = NULL, rotate = FALSE, ...)
}
\arguments{
\item{x}{SummarizedExperiment-like object with DNA methylation data}

\item{individuals}{How many distinct individuals are (allegedly) in the data?}

\item{rotate}{rotate the plot? (default is FALSE)}
}
\value{
subject identity assignments for the dataset
}
\description{
plot SNP probes from DNA methylation arrays to avoid label swaps
}

