% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotX.R
\name{plotX}
\alias{plotX}
\title{plot X chromosome probes from DNA methylation arrays to confirm gender}
\usage{
plotX(x, nFeats = 100, rotate = FALSE, ...)
}
\arguments{
\item{x}{a SummarizedExperiment-like dataset with DNA methylation data}

\item{nFeats}{how many features (max) to use for clustering (default 100)}

\item{rotate}{rotate the plot? (default is FALSE)}
}
\value{
gender assignments for the subjects
}
\description{
plot X chromosome probes from DNA methylation arrays to confirm gender
}

