% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessNoob27k.R
\name{preprocessNoob}
\alias{preprocessNoob}
\title{avoid consulting an annotation package for a platform with only Type I probes}
\usage{
preprocessNoob(rgSet, offset = 15, dyeCorr = FALSE, verbose = TRUE)
}
\arguments{
\item{rgSet}{an RGChannelSet}

\item{offset}{offset term for intensities (default is 15)}

\item{dyeCorr}{ignored (no need on 27k)}

\item{verbose}{make noise? (TRUE)}
}
\value{
a MethylSet
}
\description{
avoid consulting an annotation package for a platform with only Type I probes
}

