% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalyzeFromIDATs.R
\name{reanalyzeFromIDATs}
\alias{reanalyzeFromIDATs}
\title{as the name says, if there are IDATs attached to a grSet, rerun them.}
\usage{
reanalyzeFromIDATs(grSet, name)
}
\arguments{
\item{grSet}{a grSet with $supplementary_file and $supplementary_file.1}

\item{name}{the name of the resulting rgSet and grSet files}
}
\value{
a reanalyzed and funnorm()'ed grSet
}
\description{
as the name says, if there are IDATs attached to a grSet, rerun them.
}

