% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgSetToCNV.R
\name{rgSetToCNV}
\alias{rgSetToCNV}
\title{Take the result of IDATsToRgSet, run noob, grind out CNV calls from conumee.}
\usage{
rgSetToCNV(rgSet, cases, controls, ...)
}
\arguments{
\item{rgSet}{an RGChannelSet}

\item{cases}{which samples to run, segment, and annotate?}

\item{controls}{which samples are the control samples (for normalization)?}

\item{...}{optional arguments for conumee::CNV.create_anno function}
}
\value{
a list of CNV.analysis objects suitable for plotting etc.
}
\description{
Take the result of IDATsToRgSet, run noob, grind out CNV calls from conumee.
}

