% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgSetToGrSet.R
\name{rgSetToGrSet}
\alias{rgSetToGrSet}
\title{Take the result of IDATsToRgSet and preprocess it (pvals, SNPs, etc.)}
\usage{
rgSetToGrSet(rgSet, pcutoff = 0.01, genome = c("hg19", "hg38", "hg18"),
  funnorm = FALSE, gr = NULL, ...)
}
\arguments{
\item{rgSet}{an RGChannelSet}

\item{pcutoff}{p-value cutoff to set a beta value to NA (0.01)}

\item{genome}{genome build to annotate against if not using funnorm (hg19)}

\item{funnorm}{run preprocessFunnnorm() to get the grSet? (FALSE, use noob)}

\item{gr}{an optional GenomicRanges to annotate the rows of the result}

\item{...}{additional arguments for preprocessNoob (eg. how="reference")}
}
\value{
a GenomicRatioSet with metadata(grSet)$SNPs and p > 0.01 set to NA
}
\description{
Take the result of IDATsToRgSet and preprocess it (pvals, SNPs, etc.)
}

