% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveBy.R
\name{saveBy}
\alias{saveBy}
\alias{saveByArm}
\alias{saveByChr}
\alias{saveByCytoband}
\title{save-by-subset functions for huge (n > 1000) datasets for e.g. corpcor(x)}
\usage{
saveBy(x, y, name)

saveByChr(x, name)

saveByArm(x, name)

saveByCytoband(x, name)
}
\arguments{
\item{x}{a SummarizedExperiment-derived object or something like it}

\item{y}{how to split x before saving it into subsets via mcols(x)$y}

\item{name}{a prefix for the saved RDS files, becomes "name_subset.rds"}
}
\description{
save-by-subset functions for huge (n > 1000) datasets for e.g. corpcor(x)
}
\section{Functions}{
\itemize{
\item \code{saveByChr}: save x by chromosome

\item \code{saveByArm}: save x by chromosome arm

\item \code{saveByCytoband}: save x by cytoband
}}

