% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBy.R
\name{splitBy}
\alias{byArm}
\alias{byChr}
\alias{byCytoband}
\alias{splitBy}
\title{This might be useful for _all_ SummarizedExperiments; split x on mcols[[y]]
If the column 'y' isn't in mcols(x), it will be added by addCytoInfo(x, y)}
\usage{
splitBy(x, y = c("chr", "chromosome", "arm", "band", "cytoband"))

byChr(x)

byArm(x)

byCytoband(x)
}
\arguments{
\item{x}{a SummarizedExperiment-derived object or something like it}

\item{y}{what kind of information to split on (chr, arm, band)}
}
\value{
x, split by y
}
\description{
This might be useful for _all_ SummarizedExperiments; split x on mcols[[y]]
If the column 'y' isn't in mcols(x), it will be added by addCytoInfo(x, y)
}
\section{Functions}{
\itemize{
\item \code{byChr}: split x by chromosome

\item \code{byArm}: split x by chromosome arm

\item \code{byCytoband}: split x by cytoband
}}

