% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strpop}
\alias{strpop}
\title{pop a string off a string, for (e.g.) grabbing source barcodes from GEO}
\usage{
strpop(x, y = "_", z = NULL)
}
\arguments{
\item{x}{a string}

\item{y}{a split character}

\item{z}{which element(s) to retrieve (the last one)}
}
\value{
the element(S) produced by splitting basename(x) on y
}
\description{
pop a string off a string, for (e.g.) grabbing source barcodes from GEO
}

