% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strpush}
\alias{strpush}
\title{The opposite of strpop: push string(s) onto a string.}
\usage{
strpush(w, x, y = "_", z = NULL)
}
\arguments{
\item{w}{a string}

\item{x}{one or more elements to push onto the string}

\item{y}{optional, split character (default: "_")}

\item{z}{optional, add x after str2vec(w)[z] (default: length(str2vec(w)))}
}
\value{
a string
}
\description{
The opposite of strpop: push string(s) onto a string.
}

