% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strshift}
\alias{strshift}
\title{grab the z'th element(s) off of a vector formed by splitting x on y 
useful for (e.g.) grabbing source barcodes from GEO}
\usage{
strshift(x, y = "_", z = 1)
}
\arguments{
\item{x}{a string}

\item{y}{a split character ("_")}

\item{z}{which element(s) to retrieve (1)}
}
\value{
the element(s) [z] produced by splitting basename(x) on y
}
\description{
grab the z'th element(s) off of a vector formed by splitting x on y 
useful for (e.g.) grabbing source barcodes from GEO
}

