% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch27k.R
\name{fetch27k}
\alias{fetch27k}
\title{Pull a 27k dataset from GEO.}
\usage{
fetch27k(GSE, IDATs = TRUE, ...)
}
\arguments{
\item{GSE}{The dataset.}

\item{IDATs}{Reanalyze from IDAT files if possible? (TRUE)}

\item{...}{Additional arguments to be passed to prepForReanalysis().}
}
\value{
a GenomicRatioSet.
}
\description{
Doesn't matter what the annotations are or any other bullshit. Just get it.
}

