% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIDATsToRgSet.R
\name{geoIDATsToRgSet}
\alias{geoIDATsToRgSet}
\title{read in gzipped IDAT files corresponding to a GEO series with IDAT files}
\usage{
geoIDATsToRgSet(res, parallel = FALSE)
}
\arguments{
\item{res}{something with $supplementary_file and $supplementary_file.1}

\item{parallel}{try to run multicore (default: FALSE)}
}
\value{
an rgSet
}
\description{
read in gzipped IDAT files corresponding to a GEO series with IDAT files
}

