% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R, R/MultiAssayExperiment-methods.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{MultiAssayExperiment-class}
\alias{[,MultiAssayExperiment,ANY,ANY,ANY-method}
\title{Something like a SummarizedExperiment, but with referential integrity checks.
This is mostly a proof of concept for a well-behaved integration container.}
\usage{
\S4method{[}{MultiAssayExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a MultiAssayExperiment}

\item{i}{optional index or indices}

\item{j}{optional index or indices}
}
\value{
A valid MultiAssayExperiment w/column-matching linked assays
}
\description{
Something like a SummarizedExperiment, but with referential integrity checks.
This is mostly a proof of concept for a well-behaved integration container.
}
\details{
perSampleMetadata will shortly be renamed to "linkedAssays".

If and only if 'j' is set, the MultiAssayExperiment 'x' 
               briefly becomes a 'RangedSummarizedExperiment' while its 
               linkedAssays are recursively subsetted. The two pieces are 
               then put back together as a valid MultiAssayExperiment.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: A referential-integrity-preserving subset operator
}}
\section{Slots}{

\describe{
\item{\code{perSampleMetadata}}{associated per-sample data, checked against colnames}
}}

