% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBarcode.R
\name{parseBarcode}
\alias{parseBarcode}
\title{Parse a TCGA or TARGET barcode and translate the codebook entries.
Note that tissue source site is inconsistent between TCGA and TARGET
and is discarded by default to avoid confusion.}
\usage{
parseBarcode(barcode, sourcesite = FALSE)
}
\arguments{
\item{barcode}{a TCGA or TARGET barcode for a sample/specimen}

\item{sourcesite}{retain and translate the tissue Source.Site? (FALSE)}
}
\value{
a vector of named fragments
}
\description{
Parse a TCGA or TARGET barcode and translate the codebook entries.
Note that tissue source site is inconsistent between TCGA and TARGET
and is discarded by default to avoid confusion.
}

