% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNPs.R
\name{plotSNPs}
\alias{plotSNPs}
\title{plot SNP probes from DNA methylation arrays to avoid label swaps}
\usage{
plotSNPs(x, rotate = FALSE, ...)
}
\arguments{
\item{x}{SummarizedExperiment-like object with DNA methylation data}

\item{rotate}{rotate the plot? (default is FALSE)}
}
\value{
output from heatmap(...)
}
\description{
plot SNP probes from DNA methylation arrays to avoid label swaps
}

